20231121 fix code structure
