+++ .github/workflows/unit_test.yml
@@ -5,7 +5,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ['3.8', '3.9', '3.10', '3.11']
+        python-version: ['3.8', '3.9', '3.10', '3.11', '3.12']
     steps:
       - name: Checkout Repository
         uses: actions/checkout@v2
@@ -16,4 +16,4 @@ jobs:
       - name: Install dependencies
         run: pip install -r requirements.txt
       - name: Run Tests
-        run: python -m unittest discover -s unittest -p 'test*.py'
+        run: python -m unittest discover -s coverage/unittest -p 'UT*.py'
