+++ .gitignore
@@ -1,3 +1,6 @@
 .DS_Store
 __pycache__
+.env
+
+# tmp files from examples
 result
