+++ env
@@ -1,0 +1,9 @@
+# .env
+
+# Define valid input types for data classification
+VALID_IMAGE_TYPES = "i image"
+VALID_VIDEO_TYPES = "v video"
+
+# Define valid file extensions for images and videos
+IMAGE_EXT=".jpg .jpeg .webp .bmp .png"
+VIDEO_EXT=".mp4 .avi .mov .mkv .wmv"
